backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.9
pydantic==2.10.6
SQLAlchemy==2.0.38
psycopg2-binary==2.9.9
alembic==1.13.2
python-docx==1.1.2
pdfplumber==0.11.4
Pillow==10.4.0
pytesseract==0.3.13
google-generativeai==0.8.4
requests==2.32.3
